+++ src/main/java/ph/devcon/rapidpass/exceptions/AccessPassNotFoundException.java
@@ -1,0 +1,11 @@
+package ph.devcon.rapidpass.exceptions;
+
+/**
+ * @author czeideavanzado
+ */
+public class AccessPassNotFoundException extends Exception {
+
+    public AccessPassNotFoundException(String message) {
+        super(message);
+    }
+}